Split up Expression
