+++ tests/Components/ArrayObjTest.php
@@ -8,6 +8,7 @@ use PhpMyAdmin\SqlParser\Components\ArrayObj;
 use PhpMyAdmin\SqlParser\Components\Expression;
 use PhpMyAdmin\SqlParser\Parser;
 use PhpMyAdmin\SqlParser\Parsers\ArrayObjs;
+use PhpMyAdmin\SqlParser\Parsers\Expressions;
 use PhpMyAdmin\SqlParser\Tests\TestCase;
 use PHPUnit\Framework\Attributes\DataProvider;
 
@@ -31,7 +32,7 @@ class ArrayObjTest extends TestCase
             new Parser(),
             $this->getTokensList('(1 + 2, 3 + 4)'),
             [
-                'type' => Expression::class,
+                'type' => Expressions::class,
                 'typeOptions' => ['breakOnParentheses' => true],
             ],
         );
